Adde role control on all rest methods

+++ app-specific-code/src/main/java/org/oskari/example/st/STOperationMethodHandler.java
@@ -25,6 +25,7 @@ import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
+import org.oskari.example.UPTRoles;
 
 @OskariActionRoute("st_operation")
 public class STOperationMethodHandler extends RestActionHandler {
@@ -63,6 +64,10 @@ public class STOperationMethodHandler extends RestActionHandler {
                         stUser,
                         stPassword);) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
             PreparedStatement statement = connection.prepareStatement(
                     "SELECT id, value, language, label, created, updated\n"
                     + "FROM public.st_join_options");
@@ -87,7 +92,7 @@ public class STOperationMethodHandler extends RestActionHandler {
                 out.put(json);
             }
             ResponseHelper.writeResponse(params, out);
-        } catch (SQLException e) {
+        } catch (Exception e) {
             try {
                 errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("Error", e.toString()))));
                 ResponseHelper.writeError(params, "", 500, new JSONObject().put("Errors", errors));
@@ -98,9 +103,7 @@ public class STOperationMethodHandler extends RestActionHandler {
             }
             errorMsg = errorMsg + e.toString();
             log.error(e, errorMsg);
-        } catch (JsonProcessingException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethod.class.getName()).log(Level.SEVERE, null, ex);
-        } 
+        }
     }
 
     @Override
@@ -118,6 +121,11 @@ public class STOperationMethodHandler extends RestActionHandler {
                         stUser,
                         stPassword);) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "INSERT INTO public.st_join_options(\n"
                     + "value, language, label)\n"
@@ -133,7 +141,7 @@ public class STOperationMethodHandler extends RestActionHandler {
             errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("OK", "Operation method registered"))));
             ResponseHelper.writeResponse(params, new JSONObject().put("Errors", errors));
 
-        } catch (SQLException e) {
+        } catch (Exception e) {
             try {
                 errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("Error", e.toString()))));
                 ResponseHelper.writeError(params, "", 500, new JSONObject().put("Errors", errors));
@@ -145,10 +153,6 @@ public class STOperationMethodHandler extends RestActionHandler {
             
             errorMsg = errorMsg + e.toString();
             log.error(e, errorMsg);
-        } catch (JsonProcessingException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethodHandler.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (JSONException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethodHandler.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
@@ -168,6 +172,11 @@ public class STOperationMethodHandler extends RestActionHandler {
                         stUser,
                         stPassword);) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "UPDATE public.st_join_options\n"
                     + "SET value=?, language=?, label=?\n"
@@ -184,7 +193,7 @@ public class STOperationMethodHandler extends RestActionHandler {
 
             errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("OK", "Operation method registered"))));
             ResponseHelper.writeResponse(params, new JSONObject().put("Errors", errors));
-        } catch (SQLException e) {
+        } catch (Exception e) {
             try {
                 errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("Error", e.toString()))));
                 ResponseHelper.writeError(params, "", 500, new JSONObject().put("Errors", errors));
@@ -196,10 +205,6 @@ public class STOperationMethodHandler extends RestActionHandler {
             
             errorMsg = errorMsg + e.toString();
             log.error(e, errorMsg);
-        } catch (JsonProcessingException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethodHandler.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (JSONException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethodHandler.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
@@ -216,6 +221,11 @@ public class STOperationMethodHandler extends RestActionHandler {
                         stUser,
                         stPassword);) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM public.st_join_options\n"
                     + "WHERE id=?;");
@@ -225,7 +235,7 @@ public class STOperationMethodHandler extends RestActionHandler {
             
             statement.execute();
 
-        } catch (SQLException e) {
+        } catch (Exception e) {
             try {
                 errors.put(JSONHelper.createJSONObject(Obj.writeValueAsString(new PostStatus("Error", e.toString()))));
                 ResponseHelper.writeError(params, "", 500, new JSONObject().put("Errors", errors));                
@@ -237,8 +247,6 @@ public class STOperationMethodHandler extends RestActionHandler {
             
             errorMsg = errorMsg + e.toString();
             log.error(e, errorMsg);
-        } catch (JsonProcessingException ex) {
-            java.util.logging.Logger.getLogger(STOperationMethodHandler.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 }
