+++ app-specific-code/src/main/java/org/oskari/example/up/UPIndicatorsHandler.java
@@ -23,6 +23,8 @@ import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
+import java.util.ArrayList;
+import org.oskari.example.UPTRoles;
 
 @OskariActionRoute("indicators_manager")
 public class UPIndicatorsHandler extends RestActionHandler {
@@ -65,6 +67,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "SELECT id, indicator\n"
                     + "	FROM public.up_indicators;"
@@ -96,6 +103,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "INSERT INTO public.up_indicators(\n"
                     + "	indicator)\n"
@@ -124,6 +136,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "UPDATE public.up_indicators\n"
                     + "	SET indicator=?\n"
@@ -153,6 +170,11 @@ public class UPIndicatorsHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM public.up_indicators\n"
                     + "	WHERE id=?;"
