+++ app-specific-code/src/main/java/org/oskari/example/up/UPIndicatorsLabelingHandler.java
@@ -23,6 +23,8 @@ import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
+import java.util.ArrayList;
+import org.oskari.example.UPTRoles;
 
 @OskariActionRoute("indicators_labeling")
 public class UPIndicatorsLabelingHandler extends RestActionHandler {
@@ -59,6 +61,11 @@ public class UPIndicatorsLabelingHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "SELECT id, language, label, units, up_indicators_id\n"
                     + "	FROM public.up_indicators_translation;"
@@ -95,6 +102,11 @@ public class UPIndicatorsLabelingHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "INSERT INTO public.up_indicators_translation(\n"
                     + "	language, label, units, up_indicators_id)\n"
@@ -125,6 +137,10 @@ public class UPIndicatorsLabelingHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
 
             PreparedStatement statement = connection.prepareStatement(
                     "UPDATE public.up_indicators_translation\n"
@@ -157,6 +173,11 @@ public class UPIndicatorsLabelingHandler extends RestActionHandler {
                         upUser,
                         upPassword)) {
             params.requireLoggedInUser();
+            ArrayList<String> roles = new UPTRoles().handleGet(params,params.getUser());
+            if (!roles.contains("UPTAdmin") && !roles.contains("UPTUser") ){
+                throw new Exception("User privilege is not enough for this action");
+            }
+            
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM public.up_indicators_translation\n"
                     + "	WHERE id=?;"
